requirements.txt
boto3>=1.16.1
Django>=3.1.6
django-crispy-forms>=1.11.0
django-debug-toolbar>=3.1.1
django-extensions==3.0.9
django-heroku>=0.3.1
django-storages>=1.10.1
googlemaps>=4.4.2
googletrans>=3.0.0
gunicorn>=20.0.4
ipython>=7.19.0
psycopg2>=2.8.6
sendgrid-python>=0.1.1